requirements.txt
discord.py>=2.7.1
aiohttp>=3.13.4
requests>=2.33.1
python-dotenv>=1.2.2
python-dateutil>=2.8.2
playwright>=1.59.0
beautifulsoup4>=4.12.0
psutil>=5.9.0